1. Fixed bug： 消息ID：int-> long 2. 去除消息模块HttpSession，进一步解耦Servlet的API。

+++ shared/src/main/java/lemon/shared/message/local/persistence/LocalMsgBeanRepository.java
@@ -43,7 +43,7 @@ public interface LocalMsgBeanRepository {
 	 * @return
 	 */
 	@SelectProvider(type = LocalMsgBeanSQLProvider.class, method = "getMsgSQL")
-	LocalMsgBean getMsg(@Param("id") int id, @Param("level") int level);
+	LocalMsgBean getMsg(@Param("id") long id, @Param("level") int level);
 	
 	/**
 	 * 更新RobotMsgBean
